Refactoring of code and removed part of duplicate code
1. Made changes to return Seq[Ast] by every Ast creator function. This
will facilitate reusable methods that can follow consistent return
types.
Where ever caller expects a single value to be returned, we can always
extract the value by retrieving the head of the sequence.
2. Changed the places where empty Ast() nodes are returned and instead
returned empty Seq().
