+++ joern-cli/frontends/gosrc2cpg/src/main/scala/io/joern/gosrc2cpg/astcreation/AstCreator.scala
@@ -16,7 +16,7 @@ import ujson.Value
 class AstCreator(val relPathFileName: String, val parserResult: ParserResult)
     extends AstCreatorBase(relPathFileName)
     with AstCreatorHelper
-    with AstForDeclarationCreator
+    with AstForGenDeclarationCreator
     with AstForExpressionCreator
     with AstForFunctionsCreator
     with AstForPrimitivesCreator
@@ -64,21 +64,21 @@ class AstCreator(val relPathFileName: String, val parserResult: ParserResult)
     val blockNode_ = blockNode(rootNode, Defines.empty, Defines.anyTypeName)
 
     val methodReturn = methodReturnNode(rootNode, Defines.anyTypeName)
-    val declsAsts    = rootNode.json(ParserKeys.Decls).arr.map(item => astForNode(item)).toList
+    val declsAsts    = rootNode.json(ParserKeys.Decls).arr.flatMap(item => astForNode(item)).toList
     Ast(fakeGlobalTypeDecl).withChild(
       methodAst(fakeGlobalMethod, Seq.empty, blockAst(blockNode_, declsAsts), methodReturn)
     )
   }
 
-  protected def astForNode(json: Value): Ast = {
+  protected def astForNode(json: Value): Seq[Ast] = {
     val nodeInfo = createParserNodeInfo(json)
     val output = nodeInfo.node match {
-      case GenDecl if isImportDeclaration(nodeInfo) => astForImport(nodeInfo)
-      case BasicLit                                 => astForLiteral(nodeInfo)
-      case Ident                                    => astForIdentifier(nodeInfo)
-      case FuncDecl                                 => astForFuncDecl(nodeInfo)
-      case _: BaseExprStmt                          => astForExpression(nodeInfo)
-      case _                                        => Ast()
+      case GenDecl         => astForGenDecl(nodeInfo)
+      case BasicLit        => astForLiteral(nodeInfo)
+      case Ident           => astForIdentifier(nodeInfo)
+      case FuncDecl        => astForFuncDecl(nodeInfo)
+      case _: BaseExprStmt => astForExpression(nodeInfo)
+      case _               => Seq()
     }
     output
   }
