+++ joern-cli/frontends/gosrc2cpg/src/main/scala/io/joern/gosrc2cpg/astcreation/AstForStatementsCreator.scala
@@ -9,36 +9,31 @@ import io.shiftleft.codepropertygraph.generated.{DispatchTypes, Operators}
 import scala.util.Try
 
 trait AstForStatementsCreator { this: AstCreator =>
-  def astForBlockStatement(blockStmt: ParserNodeInfo) = {
+  def astForBlockStatement(blockStmt: ParserNodeInfo): Seq[Ast] = {
 
     val newBlockNode = blockNode(blockStmt)
     scope.pushNewScope(newBlockNode)
-    val childAsts = Try(blockStmt.json(ParserKeys.List).arr.map(createParserNodeInfo).flatMap { parserNode =>
-      parserNode.node match {
-        case DeclStmt   => astForDeclStatement(parserNode)
-        case AssignStmt => astForAssignStatement(parserNode)
-        case IncDecStmt => Seq(astForIncDecStatement(parserNode))
-        case _          => Seq(Ast())
+    val childAsts =
+      blockStmt.json(ParserKeys.List).arrOpt.getOrElse(List()).arr.map(createParserNodeInfo).flatMap { parserNode =>
+        parserNode.node match {
+          case DeclStmt   => astForDeclStatement(parserNode)
+          case AssignStmt => astForAssignStatement(parserNode)
+          case IncDecStmt => astForIncDecStatement(parserNode)
+          case _          => Seq()
+        }
       }
-    }).toOption.getOrElse(Seq(Ast()))
     scope.popScope()
-    blockAst(newBlockNode, childAsts.toList)
+    Seq(blockAst(newBlockNode, childAsts.toList))
   }
 
-  private def astForDeclStatement(declStmt: ParserNodeInfo) = {
+  private def astForDeclStatement(declStmt: ParserNodeInfo): Seq[Ast] = {
     val nodeInfo = createParserNodeInfo(declStmt.json(ParserKeys.Decl))
     nodeInfo.node match {
-      case GenDecl =>
-        nodeInfo.json(ParserKeys.Specs).arr.map(createParserNodeInfo).flatMap { parserNode =>
-          parserNode.node match {
-            case ValueSpec => astForValueSpec(parserNode)
-            case _         => Seq(Ast())
-          }
-        }
+      case GenDecl => astForGenDecl(nodeInfo)
     }
   }
 
-  private def astForAssignStatement(assignStmt: ParserNodeInfo) = {
+  private def astForAssignStatement(assignStmt: ParserNodeInfo): Seq[Ast] = {
     val op = assignStmt.json(ParserKeys.Tok).value match {
       case "="   => Operators.assignment
       case ":="  => Operators.assignment
@@ -59,17 +54,17 @@ trait AstForStatementsCreator { this: AstCreator =>
     // create corresponding local node as this is known as short variable declaration operator
     val localNodesIfIntialized =
       if (assignStmt.json(ParserKeys.Tok).value == ":=") createLocalNodeForShortVariableDeclaration(assignStmt)
-      else Ast()
-    val arguments = assignStmt.json(ParserKeys.Lhs).arr.map(astForNode).toList ::: assignStmt
+      else Seq()
+    val arguments = assignStmt.json(ParserKeys.Lhs).arr.flatMap(astForNode).toList ::: assignStmt
       .json(ParserKeys.Rhs)
       .arr
-      .map(astForNode)
+      .flatMap(astForNode)
       .toList
-    val callAst_ = callAst(cNode, arguments)
-    Seq(localNodesIfIntialized, callAst_)
+    val callAst_ = Seq(callAst(cNode, arguments))
+    localNodesIfIntialized ++: callAst_
   }
 
-  private def astForIncDecStatement(incDecStatement: ParserNodeInfo) = {
+  private def astForIncDecStatement(incDecStatement: ParserNodeInfo): Seq[Ast] = {
     val op = incDecStatement.json(ParserKeys.Tok).value match {
       case "++" => Operators.postIncrement
       case "--" => Operators.postDecrement
@@ -77,10 +72,10 @@ trait AstForStatementsCreator { this: AstCreator =>
     }
     val cNode   = callNode(incDecStatement, incDecStatement.code, op, op, DispatchTypes.STATIC_DISPATCH)
     val operand = astForNode(incDecStatement.json(ParserKeys.X))
-    callAst(cNode, Seq(operand))
+    Seq(callAst(cNode, (operand)))
   }
 
-  def createLocalNodeForShortVariableDeclaration(assignStmt: ParserNodeInfo): Ast = {
+  def createLocalNodeForShortVariableDeclaration(assignStmt: ParserNodeInfo): Seq[Ast] = {
 
     val localNodes = (assignStmt.json(ParserKeys.Lhs).arr zip assignStmt.json(ParserKeys.Rhs).arr)
       .map { case (lhs, rhs) => (createParserNodeInfo(lhs), createParserNodeInfo(rhs)) }
@@ -91,38 +86,7 @@ trait AstForStatementsCreator { this: AstCreator =>
         scope.addToScope(name, (node, typ))
         node
       }
-    Ast(localNodes)
-  }
-
-  private def astForValueSpec(valueSpec: ParserNodeInfo) = {
-
-    val localNodes = valueSpec.json(ParserKeys.Names).arr.map { parserNode =>
-      val localParserNode = createParserNodeInfo(parserNode)
-
-      val name = parserNode(ParserKeys.Name).str
-      val typ  = valueSpec.json(ParserKeys.Type).obj(ParserKeys.Name).str
-      val node = localNode(localParserNode, name, localParserNode.code, typ)
-      scope.addToScope(name, (node, typ))
-      Ast(node)
-    }
-
-    if (!valueSpec.json(ParserKeys.Values).isNull) {
-      val callNodes = (valueSpec.json(ParserKeys.Names).arr.toList zip valueSpec.json(ParserKeys.Values).arr.toList)
-        .map { case (lhs, rhs) => (createParserNodeInfo(lhs), createParserNodeInfo(rhs)) }
-        .map { case (lhsParserNode, rhsParserNode) =>
-          val cNode = callNode(
-            rhsParserNode,
-            lhsParserNode.code + rhsParserNode.code,
-            Operators.assignment,
-            Operators.assignment,
-            DispatchTypes.STATIC_DISPATCH
-          )
-          val arguments = Seq(astForNode(lhsParserNode.json), astForNode(rhsParserNode.json))
-          callAst(cNode, arguments)
-        }
-      localNodes.toList ::: callNodes
-    } else
-      localNodes.toList
+    Seq(Ast(localNodes))
   }
 
 }
